[#660] package CalculatorView
